adding a parent class for calling synthetic codes
